BigDecimal instead of Double

+++ src/main/scala/com/tvi/app/SCPP.scala
@@ -10,7 +10,7 @@ package com.tvi.app
 case class SCPP(customerId: String,
                 startTime: String,
                 endTime: String,
-                volume: Double)
+                volume: BigDecimal)
     extends CustomToString
 
 /**
